feat: mouse hover show hidden filename and change preview screen
