+++ src/components/CourseSearchBox.css
@@ -18,76 +18,3 @@
   justify-content: center;
   align-items: center;
 }
-
-.itemBox {
-  list-style: none;
-  width: 100%;
-  border-bottom: 1px solid #ccc;
-  padding: 10px;
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  align-items: center;
-  font-weight: bold;
-  color: #000;
-}
-
-/* .itemText {
-  height: 30px;
-  border-radius: 3px;
-  display: flex;
-  width: 280px;
-  overflow: hidden;
-  padding: 5px;
-  align-items: center;
-}
-
-.itemText span {
-  width: 280px;
-  color: #000;
-  text-overflow: ellipsis;
-  align-items: center;
-  overflow: hidden;
-  white-space: nowrap;
-} */
-
-.itemText {
-  overflow: hidden;
-  white-space: nowrap;
-}
-
-.itemText span {
-  display: inline-block;
-  will-change: transform;
-  transition: transform 1s linear;
-}
-
-.itemText:hover span {
-  transform: translateX(-100%);
-}
-
-.elapsedTime {
-  display: flex;
-  font-size: 12px;
-  font-weight: 400;
-  background-color: #999;
-  border-radius: 10px;
-  color: white;
-  width: 100%;
-  padding-left: 8px;
-  padding-right: 8px;
-  text-align: center;
-  justify-content: center;
-  align-items: center;
-  cursor: default;
-}
-
-.leftpanel {
-  display: flex;
-  width: 280px;
-}
-
-.rightpanel {
-  display: flex;
-  flex-direction: row;
-}
